Reorder PSD Constraint Rows for Clarabel
Diffcp requires that the `A` matrix conforms to the SCS format,
including that PSD constraints refer to the columnwise stacking of the
lower triangle. Clarabel expects that the PSD constraints refer to the
upper triangle, leading to an incorrect result when using Clarabel with
PSD constraints.

A unit test showing the issue was added to test_clarabel. In this test,
X is a 3x3 PSD matrix and we wish to minimize X00 - X22 while
constraining the trace to be equal to 1. Before this change, Clarabel
reports success, but the resulting matrix isn't even PSD. After this
change, we get the expected result.